Conflicts between the cytoverse and other packages

This function lists all the conflicts between packages in the cytoverse and other packages that you have loaded.

cytoverse_conflicts()

Details

There are four conflicts that are deliberately ignored: intersect, union, setequal, and setdiff from dplyr. These functions make the base equivalents generic, so shouldn't negatively affect any existing code.

Examples

cytoverse_conflicts()

#> ── Conflicts ─────────────────────────────────────────────────────────────────── cytoverse_conflicts() ──
#> ✖ ggcyto::%+%()               masks ggplot2::%+%()
#> ✖ flowCore::%in%()            masks base::%in%()
#> ✖ flowCore::alias()           masks stats::alias()
#> ✖ flowCore::assign()          masks base::assign()
#> ✖ CytoML::colnames()          masks flowWorkspace::colnames(), ncdfFlow::colnames(), flowCore::colnames(), base::colnames()
#> ✖ flowWorkspace::colnames<-() masks ncdfFlow::colnames<-(), flowCore::colnames<-(), base::colnames<-()
#> ✖ flowCore::eval()            masks base::eval()
#> ✖ ncdfFlow::filter()          masks flowCore::filter(), stats::filter()
#> ✖ flowCore::head()            masks utils::head()
#> ✖ flowWorkspace::lapply()     masks ncdfFlow::lapply(), base::lapply()
#> ✖ flowCore::ls()              masks base::ls()
#> ✖ flowCore::ncol()            masks base::ncol()
#> ✖ flowCore::nrow()            masks base::nrow()
#> ✖ openCyto::plot()            masks flowWorkspace::plot(), flowCore::plot(), graphics::plot()
#> ✖ ggcyto::print()             masks flowCore::print(), base::print()
#> ✖ ncdfFlow::split()           masks flowCore::split(), base::split()
#> ✖ flowCore::summary()         masks base::summary()
#> ✖ flowCore::tail()            masks utils::tail()
#> ✖ ggcyto::transform()         masks flowWorkspace::transform(), ncdfFlow::transform(), flowCore::transform(), base::transform()
#> ✖ ncdfFlow::unlink()          masks base::unlink()